+++ changelog.md
@@ -6,6 +6,7 @@
 
 - Added option to synchronize directories with files on every supported protocol
 - Added support for the plugins CatFolders and Filebird to import files in specific folders of this plugins
+- Added support for the plugin Enhanced Media Library to add imported files to their categories
 - Added wrapper for all settings of this plugin for easier management in the future
 - New option for import local files from chosen directory via queue (for very large directories)